try(ProdDetail Questions): trying to render questions
