+++ PF_Client/src/components/Detail/Detail.jsx
@@ -3,9 +3,7 @@ import {
   useDisclosure,
   Button,
   Modal,
-  ModalOverlay,
   ModalContent,
-  ModalHeader,
   ModalFooter,
   ModalBody,
   ModalCloseButton,
