+++ PF_Client/src/components/Questions/Questions.module.css
@@ -1,0 +1,4 @@
+.detailTable-wrapper {
+    display: flex;
+    justify-content: center;
+  }
