+++ PF_Client/src/views/ProductDetail/ProductDetail.module.css
@@ -2,10 +2,7 @@
     color: aqua;
 }
 
-.detailTable-wrapper {
-    display: flex;
-    justify-content: center;
-  }
+
 
 .bg{
     background-repeat: no-repeat;
